requirements.txt
streamlit 
openai 
pdfplumber 
pandas
python-dotenv
